Do not retry in XLogPageRead() after crash recovery
In upstream commit 1c7fe7db, retry was added in XLogPageRead() when the second
page header of a continuation record is invalid on standby, so that standby can
receive an up-to-date good page later on from streaming.

However, this retry logic causes startup process to stuck in
WaitForWALToBecomeAvailable() if this continuation record is discarded on
master after crash recovery and partially exists in xlogs on mirror.

This patch removes the retry if a crash recovery just happend.

WIP: Move break condition inside of WaitForWALToBecomeAvailable()

Co-authored-by: Alexandra Wang <lewang@pivotal.io>
Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>