Set relpersistence when adding partition
makeRangeVar is a function from upstream which is basically a
constructor. It sets relpersistence in the RangeVar to a hard-coded
value of RELPERSISTENCE_PERMANENT. This may or may not cause problems
upstream (TODO: look for a case where this causes problems upstream).
However, in GPDB, we use the root partition to construct the RangeVar
that eventually is used to make the pg_class entry for child partitions
in multiple places.

This commit specifically only sets it back for the case in which we are
adding a partition with `ADD PARTITION` or through `SPLIT PARTITION`.

Without this commit, a leaf partition of an unlogged table created
through `ADD PARTITION` or `SPLIT PARTITION` would incorrectly have its
relpersistence set to permanent.

It is probably wrong in most cases where we are using makeRangeVar to
reset it to permanent, as we often pass use it when we have an existing
relation with its own pg_class entry which we could use to set the
relpersistence.