SET DISTRIBUTED BY: distribution keys must be prefix of unique indexes
Previously when ALTER TABLE SET DISTRIBUTED BY, we only verify that
DISTRIBUTED BY columns must be equal to or a subset of the UNIQUE INDEX
columns. Theoretically there is nothing wrong with it, however, given
the fact that CREATE TABLE and ALTER TABLE ADD PARTITION have always
been limiting the DISTRIBUTED BY columns to be equal to or a
left-subset(prefix) of the UNIQUE INDEX columns, we have to apply the
same restriction to SET DISTRIBUTED BY. Otherwise, it is possible to
successfully set DISTRIBUTED BY columns to be a subset but not a
left-subset of the UNIQUE INDEX columns, but after this new partitions
can never be added any more.

Co-authored-by: Jinbao Chen <jinchen@pivotal.io>