Add unique index: distribution keys must be prefix of unique indexes
Previously when ALTER TABLE ADD UNIQUE INDEX, we only verify that
DISTRIBUTED BY columns must be equal to or a subset of the UNIQUE INDEX
columns. Theoretically there is nothing wrong with it, however, given
the fact that CREATE TABLE and ALTER TABLE ADD PARTITION have always
been limiting the DISTRIBUTED BY columns to be equal to or a
left-subset(prefix) of the UNIQUE INDEX columns, we have to apply the
same restriction to ADD UNIQUE INDEX. Otherwise, it is possible to
successfully add an unique index that contains all the distributed by
columns but with different order, after this new partitions can never be
added.