WIP: Fix deadlock between gprecoverseg and FTS probe
The deadlock happens when gprecoverseg invokes
gp_{add,remove}_segment_mirror(). Current sequence of opterations for
those two functions are:

1. Lock gp_segment_configuration table with `AccessExclusiveLock`
2. Read the dbid of the mirror and its primary from the table
3. Acquire `RowExclusiveLock` on gp_segment_configuration
4. Add/Remove the mirror entry from gp_segment_configuration
5. Close gp_segment_configuration but keep the locks
6. Add/Remove entries in the psersistent tables that refers to the
mirror on master, and dispatch this operation to all the segments
7. Commit and release the locks on gp_segment_configuration

In step 6, the dispatcher process polls results from the QEs, and if it
encounters internal error or timeouts during the poll(), it notifies the
FTS process to probe the segments. This is where deadlock happens. After
being notified, the FTS process, specifically in getCdbComponentInfo(),
needs to accquire a `AccessShareLock`, but it is blocked by the
`AccessExclusiveLock` the dispacther accquired in step 1, so FTS cannot
probe, hence step 6 hangs.

This patch makes three changes to the dispatcher and FTS:

1. Downgrade the lock in step 1 to `ExclusiveLock`.

The reason why we chose AccessExclusiveLock is not clear from the commit
history, but we think the rationale is to prevent any writes to
gp_segment_configuration from the time it reads the dbid of the primary
and the mirror until the end of the transaction. Note that
RowExclusiveLock doesn't conflict with itself.

2. Upgrade the lock in FTS getCdbComponentInfo() to `RowShareLock` and
hold the lock for the entire FTS probe.

What FTS does is essentially select for update, in the sense that it
first reads from gp_segment_configuration into cdb_component_dbs in
memory, probe each segment by open connections to them, and finally
update gp_segment_configuration with `RowExclusiveLock`. We want
`RowShareLock` because we don't want the table to be modified during one
probe.

3. Don't request FTS probe if the dispatcher already held
`ExclusiveLock` or stronger

If we just do 1 and 2 to make things right, we would still have the
deadlock problem, as `ExclusiveLock` and `RowShareLock` doesn't conflict
with each other. Hence adding the check to prevent FTS probe if the
process that requested the probe already held a `ExclusiveLock` or
stronger.

Partial cherry-picked upstream commit b04aeb0a for LockHeldByMe() and
CheckRelationLockedByMe().