Fix deadlock between gprecoverseg and FTS probe
The deadlock happens when gprecoverseg invokes
gp_{add,remove}_segment_mirror(). Current sequence of opterations for
those two functions are:

1. Lock gp_segment_configuration table with `AccessExclusiveLock`
2. Read the dbid of the mirror and its primary from the table
3. Acquire `RowExclusiveLock` on gp_segment_configuration
4. Add/Remove the mirror entry from gp_segment_configuration
5. Close gp_segment_configuration but keep the locks
6. Add/Remove entries in the psersistent tables that refers to the
mirror on master, and dispatch this operation to all the segments
7. Commit and release the locks on gp_segment_configuration

In step 6, the dispatcher process polls results from the QEs, and if it
encounters internal error or timeouts during the poll(), it notifies the
FTS process to probe the segments. This is where deadlock happens. After
being notified, the FTS process, specifically in getCdbComponentInfo(),
needs to accquire a `AccessShareLock`, but it is blocked by the
`AccessExclusiveLock` the dispacther accquired in step 1, so FTS cannot
probe, hence step 6 hangs.

This patch fix the problem by downgrading the lock in step 1 to
ExclusiveLock.