Correctly report instrumentation of zero-output node
We used to incorrectly report "never executed" for a node that returns 0
rows from every segment. This was misleading because "never executed"
should indicate that the node was never executed by its parent.

A more subtle artifact of the bug is that the average count of rows was
higher than what it should be, because we only averaged among the
segments that return non-zero rows.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>