Keep the Append node for one partition scans
* Dynamic partition pruning
Even if one partition is present in the statically pruned plan, we could
still dynamically prune it away. This needs an Append node.

* Col mappings issues
When the first selected child partition's cols have different types/order
than the root partition, we can no longer re-use the colrefs of the root
partition, since colrefs are immutable. Thus, we create new colrefs for
this partition. But, if there is no Append (in case of just one selected
partition), then we also go through update all references above the DTS
with the new colrefs. For simplicity, we decided to keep the Append
around to maintain this projection (mapping) from the old root colrefs
to the first selected partition colrefs.