Improve ORCA partition pruning for OR'd equal comparisons on part key
Previously, a predicate that contains only disjunction(s) of equal
comparisons, with one side being the partition key, such as:

"partkey = 1 OR partkey = 2"
"partkey IN ('a', 'b')"

is the partition selector's general filter. A general filter is always
evaluated with every partition rule of the current partition level. For
a range partitioned table with many partitions, iterating through every
partition rule for just a few OR'd equal comparisons is inefficient
comparing to performing binary searches with the few given values. We
also always conservatively select the default partition for all general
filters, for both range partitions and list partitions.

This patch improves partition selection for this type of predicates by
re-categorizing them as equal filters. To do this, the following changes
are made:

1. CTranslatorExprToDXL
Examine the general filter of the CPhysicalPartitionSelector, if it
contains only disjunction(s) of equality comparisons, add them into the
list of equality filter elements of the equal filter CDXLNode instead of
the general filter node. This introduces DXL plan changes as below for
predicate "where b in (0,1)"

Before:
<dxl:PartEqFilters>
  <dxl:ConstValue TypeMdid="0.16.1.0" Value="true"/>
</dxl:PartEqFilters>
<dxl:PartFilters>
  <dxl:Or>
    <dxl:Or>
      <dxl:ArrayComp OperatorName="=" OperatorMdid="0.96.1.0" OperatorType="Any">
        <dxl:ConstValue TypeMdid="0.23.1.0" Value="0"/>
        <dxl:PartListValues Level="0" ResultType="0.1007.1.0" ElementType="0.23.1.0"/>
      </dxl:ArrayComp>
      <dxl:DefaultPart Level="0"/>
    </dxl:Or>
    <dxl:Or>
      <dxl:ArrayComp OperatorName="=" OperatorMdid="0.96.1.0" OperatorType="Any">
        <dxl:ConstValue TypeMdid="0.23.1.0" Value="1"/>
        <dxl:PartListValues Level="0" ResultType="0.1007.1.0" ElementType="0.23.1.0"/>
      </dxl:ArrayComp>
      <dxl:DefaultPart Level="0"/>
    </dxl:Or>
  </dxl:Or>
  <dxl:ConstValue TypeMdid="0.16.1.0" Value="true"/>
</dxl:PartFilters>

After:
<dxl:PartEqFilters>
  <dxl:PartEqFilterElems>
    <dxl:ConstValue TypeMdid="0.23.1.0" Value="0"/>
    <dxl:ConstValue TypeMdid="0.23.1.0" Value="1"/>
  </dxl:PartEqFilterElems>
</dxl:PartEqFilters>
<dxl:PartFilters>
  <dxl:ConstValue TypeMdid="0.16.1.0" Value="true"/>
</dxl:PartFilters>

As shown above, we encapsulate a list of scalar values for each
partition level in the newly introduced `PartEqFilterElems` DXLNode.
Most of the MDP changes in this patch are simply update the existing
non-empty[^1] `PartEqFilters` by surrounding the single scalar value
with `PartEqFilterElems`. For the interesting changes that have more
than one scalar values inside `PartEqFilterElems`, please see the
following MDP tests:

PartTbl-ArrayCoerce.mdp
PartTbl-List-DPE-Int-Predicates.mdp
PartTbl-List-DPE-Varchar-Predicates.mdp

[^1] An empty `PartEqFilters` or `PartFilters` contains one child node:
<dxl:ConstValue TypeMdid="0.16.1.0" Value="true"/>

2. CTranslatorDXLToPlStmt & Executor
`PartitionSelector->levelEqExpressions` used to be a list of scalar
values, where the list length is the number of partition levels. It is
now a list of list of scalar values, where the inner list length is the
number of OR'd equal comparisons, and the outer list length is as before
the number of partition levels. During execution, each value from the
inner list is evaluated, and the chosen partition rules for all the
values are concatenated.