Resurrect PdxlnDynamicIndexScan
Construct an Append node for DynamicIndexScan that contains a list of
IndexScans.

breadcrumbs for next week:

Map the root partitioned index to a specific child partition index.
Currently we are assuming the child partition only has one index and
just use that one.

Specifically:
Index relations are just child partitions of the partitioned index.
RetrieveRelPartitions() could be the function to get a list of leaf
indexes from the partitioned index.
Once we get that list, add that to CMDIndex.

other next steps:
- bitmap index scans
- joins

Co-authored-by: Chris Hajas <chajas@vmware.com>